+++ src/App.tsx
@@ -2,6 +2,7 @@ import {AppShell} from '@components/layout/AppShell'
 import {usePowerUpsRuntime} from '@hooks/usePowerUpsRuntime'
 import {useTauriMenu} from '@hooks/useTauriMenu'
 import {AppProvider} from '@store/context'
+import '@actions/coreActions'
 
 function AppInner() {
     useTauriMenu()